refactor(add): split add kind registry
