+++ .sampo/changesets/819-split-add-kind-registry.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split add-kind registry entries into focused per-kind modules while preserving the public aggregation point.
